+++ src/r_renderer.h
@@ -49,7 +49,7 @@ struct FRenderer
 	virtual void OnModeSet () {}
 	virtual void SetClearColor(int color) = 0;
 	virtual void Init() = 0;
-	virtual void RenderTextureView (FCanvasTexture *tex, AActor *viewpoint, int fov) = 0;
+	virtual void RenderTextureView (FCanvasTexture *tex, AActor *viewpoint, double fov) = 0;
 	virtual void PreprocessLevel() {}
 	virtual void CleanLevelData() {}
 	virtual bool RequireGLNodes() { return false; }
