+++ src/swrenderer/r_swrenderer.h
@@ -30,7 +30,7 @@ struct FSoftwareRenderer : public FRenderer
 	void OnModeSet() override;
 	void SetClearColor(int color) override;
 	void Init() override;
-	void RenderTextureView (FCanvasTexture *tex, AActor *viewpoint, int fov) override;
+	void RenderTextureView (FCanvasTexture *tex, AActor *viewpoint, double fov) override;
 
 	void PreprocessLevel() override;
 	void CleanLevelData() override;
